# docs/swagger.yaml
definitions:
  collectionmodels.ItemList:
    properties:
      base_price:
        type: number
      itemId:
        type: string
      productId:
        type: string
      quantity:
        type: integer
      sale_price:
        type: number
      sellerSku:
        type: string
      vat:
        type: number
    type: object
  collectionmodels.OrderAddress:
    properties:
      address1:
        type: string
      address2:
        type: string
      address3:
        type: string
      address4:
        type: string
      address5:
        type: string
      city:
        type: string
      country:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
      phone2:
        type: string
      post_code:
        type: string
    type: object
...
  models.ResponseAPI-any:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  models.ResponseAPI-array_models_SaleOrderReadO:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models.SaleOrderReadO'
        type: array
      message:
        type: string
    type: object
...
  models.ResponseAPI-models_SaleOrderReadO:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/models.SaleOrderReadO'
      message:
        type: string
    type: object
...
  models.SaleOrderCreateI:
    properties:
      item_list:
        description: สินค้าที่ขาย
        items:
          $ref: '#/definitions/collectionmodels.ItemList'
        type: array
      order_address:
        allOf:
        - $ref: '#/definitions/collectionmodels.OrderAddress'
        description: ที่อยู่ผู้ซื้อ
      sale_channel:
        description: ช่องทางการขาย shopee laz lineOA
        type: string
      sale_type:
        description: ประเภทการขาย โอน ปลายทาง
        type: string
      shipping_fee:
        description: ค่าขนส่ง
        type: number
      shipping_platform:
        description: ผู้จัดส่ง kerry flash
        type: string
      shop_id:
        description: รหัสร้านค้า
        type: string
      tracking_code:
        description: เลข tracking
        type: string
    type: object
  models.SaleOrderReadI:
    properties:
      orderNo:
        type: string
    type: object
  models.SaleOrderReadO:
    properties:
      item_list:
        description: สินค้าที่ขาย
        items:
          $ref: '#/definitions/collectionmodels.ItemList'
        type: array
      order_address:
        allOf:
        - $ref: '#/definitions/collectionmodels.OrderAddress'
        description: ที่อยู่ผู้ซื้อ
      order_no:
        type: string
      sale_channel:
        description: ช่องทางการขาย shopee laz lineOA
        type: string
      sale_type:
        description: ประเภทการขาย โอน ปลายทาง
        type: string
      shipping_fee:
        description: ค่าขนส่ง
        type: number
      shipping_platform:
        description: ผู้จัดส่ง kerry flash
        type: string
      shop_id:
        description: รหัสร้านค้า
        type: string
      status:
        description: สถานะ order
        type: string
      total_base_price:
        description: ต้นทุน
        type: number
      total_profit:
        description: กำไร
        type: number
      total_sale_price:
        description: ราคาที่ขาย
        type: number
      total_vat:
        description: จำนวนภาษี
        type: number
      tracking_code:
        description: เลข tracking
        type: string
    type: object
  models.SaleOrderUpdateI:
    properties:
      order_no:
        type: string
      status:
        description: สถานะ order
        type: string
      tracking_code:
        description: เลข tracking
        type: string
    type: object
...
info:
  contact: {}
  description: This is a docs of Go Backend POS API
  title: Go Backend POS API
  version: "1.0"
paths:
...
  /api/v1/saleorder:
    get:
      description: to read sale order
      operationId: SaleOrderRead
      parameters:
      - description: SaleOrderReadI
        in: body
        name: SaleOrderReadI
        schema:
          $ref: '#/definitions/models.SaleOrderReadI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-array_models_SaleOrderReadO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: SaleOrderRead
      tags:
      - SaleOrder
    post:
      description: to create sale order
      operationId: SaleOrderCreate
      parameters:
      - description: SaleOrderCreateI
        in: body
        name: SaleOrderCreateI
        schema:
          $ref: '#/definitions/models.SaleOrderCreateI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: SaleOrderCreate
      tags:
      - SaleOrder
  /api/v1/saleorder/:orderno:
    get:
      description: to Update sale order
      operationId: SaleOrderUpdate
      parameters:
      - description: SaleOrderUpdateI
        in: body
        name: SaleOrderUpdateI
        schema:
          $ref: '#/definitions/models.SaleOrderUpdateI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: SaleOrderUpdate
      tags:
      - SaleOrder
  /api/v1/saleorder/{orderno}:
    get:
      description: to read sale order by order number
      operationId: SaleOrderReadByOrderNo
      parameters:
      - description: orderno
        in: path
        name: orderno
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-models_SaleOrderReadO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: SaleOrderReadByOrderNo
      tags:
      - SaleOrder
securityDefinitions:
  bearerAuth:
    description: 'Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"